+++ src/clients/index.ts
@@ -5,3 +5,6 @@ export * from './janus'
 export * from './IOClient'
 export * from './IOClients'
 export * from './IOGraphQLClient'
+
+import promclient from 'prom-client'
+export { promclient }