+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [6.45.0] - 2021-08-24
+
+### Changed
+
+- Export Prometheus client
+
 ## [6.44.0] - 2021-07-15
 
 ### Added